Trim update (#1)
* trim script update testing

* Delete barcode09.trim.out.fastq.gz

* Delete SureSelectXT.fasta

* Update trim.py

I think I am on the final version of this script. Will get final approval, may update how `read_info` is logged, but other than that it appears to be working as desired.

* Delete barcode_locs.txt

* Delete barcode09.trim.log

* pushing all changes, should be able to merge soon

* Moved new trim script into pipeline

Moved the new trim script into the pipeline and updated it to remove
the `Strainline` steps and for the changed made to the trim script.
Will test on Hipergator before pushing to main.

* test files

* Update Snakefile

Fixed spacing and '-' in commands.

* Update trimming.yaml

* Update find_viral_targets.py

added log info to script

* Create start_counts.sh

* Update summarize_idxstats.py

added the `strain_db` as an input requirement so we can include the virus names in the outfile

* Create NOTES.md

Some notes for handling the RVHaplo output

* Update Snakefile

handles the additions for the additional output and input requirements, as well as some additional rules to get count information

* Additional testing

* Fix reads.per.strain spacing

Adding the `\t` delim and stripping end line characters should help fix the spacing issues of the reads.per.strain files.

* parse rvhaplo haplotypes into table

* Close to final updates

Added the virus names to the rvhaplo output table, added a check for empty viral target fasta, and addressed the duplicate reads issue when using `bedtools bamtofastq`.