Fixed naming of downloaded host ref file
Changed the naming so that the downloaded reference actually has the name of the organism so a user can download more than one reference. Will probably also make a check so that the workflow will check the `resources/downloaded_ref_files/` directory for a host file to see if it already exists to prevent downloading the same file multiple times.